+++ src/get-model.ts
@@ -1,14 +1,14 @@
-import { Connection, connection as defaultConnection, Model } from 'mongoose';
+import { Connection, connection as defaultConnection, Document, Model } from 'mongoose';
 import { RestRegistry } from './rest-registry';
 import { RestgooseModel } from './restgoose-model';
-import { Constructor, InstanceType } from './types';
+import { Constructor } from './types';
 
 /**
  * Get or builds the model for a specific connection
  * @param connection
  * @param model
  */
-export function getModel<T extends RestgooseModel>(model: Constructor<T>, connection?: Connection): Model<InstanceType<T>> {
+export function getModel<T extends RestgooseModel>(model: Constructor<T>, connection?: Connection): Model<T & Document> {
     if (!connection) {
         connection = defaultConnection;
     }
@@ -18,7 +18,7 @@ export function getModel<T extends RestgooseModel>(model: Constructor<T>, connec
     if (!connection.models[model.name]) {
         // get schema of current model
         const schema = model.prototype.buildSchema(schemaOptions);
-        const newModel: Model<InstanceType<T>> = connection.model(model.name, schema);
+        const newModel: Model<T & Document> = connection.model(model.name, schema);
         newModel.init();
         newModel.ensureIndexes();
         return newModel;