+++ src/middleware-composers.ts
@@ -1,7 +1,8 @@
 import { Request } from 'express';
 import { RestgooseModel } from './restgoose-model';
-import { Middleware, MiddlewarePostFetch } from './types';
-import { InstanceType } from './types';
+import { MiddlewarePostFetch, MiddlewarePreFetch, MiddlewarePreSave  } from './types';
+
+type MiddlewarePrePost<T extends RestgooseModel> = MiddlewarePreFetch | MiddlewarePostFetch<T> | MiddlewarePreSave<T>;
 
 /**
  * Compose several middlewares with a logical OR operation.
@@ -9,17 +10,22 @@ import { InstanceType } from './types';
  * Otherwise, the returned value is passed through.
  * If all the middlewares are rejected, the error thrown from the last one will be passed through.
  */
-export function or<T extends RestgooseModel, F extends Middleware>(...fns: F[]): F {
-    return ((req: Request, entity: T | boolean = true, oldEntity?: T): Promise<InstanceType<T>> => {
-        let promises: Promise<InstanceType<T>> = Promise.reject(null);
-        fns.forEach(fn => {
-            promises = promises.then(
-                v => v ? v : fn(req, entity, oldEntity),
-                () => fn(req, entity, oldEntity),
-            );
-        });
-        return promises;
-    }) as F;
+export function or<T extends RestgooseModel>(...fns: MiddlewarePrePost<T>[]): MiddlewarePrePost<T> {
+    return async (req: Request, entity: T = null, oldEntity?: T): Promise<T> => {
+        let lastError = null;
+
+        for (const fn of fns) {
+            try {
+                const out = (await fn(req, entity, oldEntity)) as T;
+                return Promise.resolve(out);
+            }
+            catch (e) {
+                lastError = e;
+            }
+        }
+
+        throw lastError;
+    };
 }
 
 /**
@@ -27,16 +33,16 @@ export function or<T extends RestgooseModel, F extends Middleware>(...fns: F[]):
  * All middlewares must pass for the entity to be returned.
  * If any middleware is rejected, the error thrown is passed through.
  */
-export function and<T extends RestgooseModel, F extends Middleware>(...fns: F[]): F {
-    return ((req: Request, entity: T | boolean = true, oldEntity?: T): Promise<InstanceType<T>> => {
+export function and<T extends RestgooseModel>(...fns: MiddlewarePrePost<T>[]): MiddlewarePrePost<T> {
+    return ((req: Request, entity: T & Document = null, oldEntity?: T & Document): Promise<T & Document> => {
         let promises: Promise<any> = Promise.resolve(entity);
         fns.forEach(m => {
             promises = promises.then(entity => {
-                return entity && m(req, entity, oldEntity);
+                return m(req, entity, oldEntity);
             });
         });
         return promises;
-    }) as F;
+    }) as MiddlewarePrePost<T>;
 }
 
 /**
@@ -45,8 +51,8 @@ export function and<T extends RestgooseModel, F extends Middleware>(...fns: F[])
  * Use it  with the 'all' method so you can use one middleware for both getting all the items
  * (with asFilter) and getting only one (without it, throwing errors).
  */
-export function asFilter<T extends RestgooseModel>(fn: MiddlewarePostFetch<T>): MiddlewarePostFetch<T> {
-    return (req: Request, entity: T, oldEntity?: T): Promise<InstanceType<T>> => {
+export function asFilter<T extends RestgooseModel>(fn: MiddlewarePrePost<T>): MiddlewarePrePost<T> {
+    return (req: Request, entity?: T & Document, oldEntity?: T & Document): Promise<T & Document> => {
         return Promise.resolve()
         .then(() => fn(req, entity, oldEntity))
         .catch(() => {