+++ src/request-util.ts
@@ -1,9 +1,8 @@
-import { Model } from 'mongoose';
+import { Document, Model } from 'mongoose';
 import { RestgooseModel } from './restgoose-model';
 import { RestRequest } from './types';
-import { InstanceType } from './types';
 
-export function buildPayload<T extends RestgooseModel>(req: RestRequest, modelType: Model<InstanceType<T>>) {
+export function buildPayload<T extends RestgooseModel>(req: RestRequest, modelType: Model<T & Document>) {
     const payload = {};
     // FIXME as any
     const properties = Object.keys((modelType.schema as any).tree);