+++ src/rest-controller.ts
@@ -1,5 +1,5 @@
 import { Response } from 'express';
-import { CastError, Types } from 'mongoose';
+import { CastError, Document, Types } from 'mongoose';
 import { debug } from './debug';
 import { ArrayPropConfiguration } from './decorators/array-prop';
 import { RestConfigurationMethod, RestError } from './decorators/rest';
@@ -13,7 +13,7 @@ import { parseQuery } from './parse-query';
 import { buildPayload } from './request-util';
 import { RestModelEntry, RestPropEntry } from './rest-registry';
 import { RestgooseModel } from './restgoose-model';
-import { Constructor, Dic, InstanceType, RestRequest } from './types';
+import { Constructor, Dic, RestRequest } from './types';
 
 export const ERROR_FORBIDDEN_CODE: string = 'FORBIDDEN';
 export const ERROR_NOT_FOUND_CODE: string = 'NOT_FOUND';
@@ -138,7 +138,7 @@ export function create<T extends RestgooseModel>(modelEntry: RestModelEntry<T>, 
             const preSaveResult = await preSave(methodConfig, req, prev, postFetchResult);
 
             // save
-            const saveResult = await persistSave(methodConfig, preSaveResult);
+            const saveResult = await persistSave(methodConfig, req, preSaveResult);
 
             // preSend
             const preSendResult = await preSend(methodConfig, req, saveResult);
@@ -199,16 +199,16 @@ export function createWithin<T extends RestgooseModel, S extends RestgooseModel>
                 const preSaveSubResult = await preSave(submethodConfig, req, prev, postFetchSubResult);
 
                 // save - sub
-                saveSubResult = await persistSave(submethodConfig, preSaveSubResult);
+                saveSubResult = await persistSave(submethodConfig, req, preSaveSubResult);
 
                 // save - parent
                 postFetchParentResult[propEntry.name].push(saveSubResult._id);
-                await persistSave(methodConfig, postFetchParentResult);
+                await persistSave(methodConfig, req, postFetchParentResult);
             }
             else {
                 // save - parent
                 postFetchParentResult[propEntry.name].push(req.body);
-                const parentSaveResult = await persistSave(methodConfig, postFetchParentResult);
+                const parentSaveResult = await persistSave(methodConfig, req, postFetchParentResult);
                 saveSubResult = parentSaveResult[propEntry.name][parentSaveResult[propEntry.name].length - 1];
             }
 
@@ -329,7 +329,7 @@ export function remove<T extends RestgooseModel>(modelEntry: RestModelEntry<T>, 
             await preSave(methodConfig, req, postFetchResult, null);
 
             // save
-            await persistDeleteOne(modelType, methodConfig, postFetchResult);
+            await persistDeleteOne(modelType, methodConfig, req, postFetchResult);
 
             return res.status(204).end();
         }
@@ -356,7 +356,7 @@ export function removeAll<T extends RestgooseModel>(modelEntry: RestModelEntry<T
         await preSaveAll(methodConfig, req, postFetchResult, new Array(postFetchResult.length).fill(null));
 
         // save
-        await persistDeleteAll(modelType, methodConfig, postFetchResult);
+        await persistDeleteAll(modelType, methodConfig, req, postFetchResult);
 
         return res.status(204).end();
     });
@@ -395,7 +395,7 @@ export function update<T extends RestgooseModel>(modelEntry: RestModelEntry<T>, 
             const preSaveResult = await preSave(methodConfig, req, prev, postFetchResult);
 
             // save
-            const saveResult = await persistSave(methodConfig, preSaveResult);
+            const saveResult = await persistSave(methodConfig, req, preSaveResult);
 
             // preSend
             const preSendResult = await preSend(methodConfig, req, saveResult);
@@ -408,7 +408,7 @@ export function update<T extends RestgooseModel>(modelEntry: RestModelEntry<T>, 
 /**
  * Deeply updates a mongoose document with a JSON object
  */
-function updateDocument<T extends RestgooseModel>(entity: InstanceType<T>, payload: Dic) {
+function updateDocument<T extends RestgooseModel>(entity: T & Document, payload: Dic) {
     for (const key in payload) {
         if (entity[key] instanceof Types.Subdocument) {
             updateDocument(entity[key], payload[key]);