+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.1 - 2019-07-26
+- `persist` hook now takes the request as argment, like the other hooks. 
+
+## 1.1.0 - 2019-06-13
+**BREAKING CHANGE** 
+Unified middlewares `preFetch`, `postFetch`, `preSave`, `preSend`. 
+Now, all middlewares can be written as `function <T extends RestgooseModel>(req: Request, entity: T): Promise<T>`.
+
+`preFetch` middlewares should now return `null` instead of `true`.
+
 ## 1.0.1 - 2019-07-01
 This version only updates development tools. No need to update.
 - Updated dependencies : codecov, mocha & tslint.